modules/connectionService/requirements.txt
grpcio==1.38.0
protobuf==3.17.3
GeoAlchemy2==0.8.4
psycopg2-binary==2.8.5
shapely==1.7.0
SQLAlchemy==1.3.19
